2018/05/18
카메라 연출 끝, 공이 부딛히는 소리 면서 거슬리지 않는정도의 소리만들기, 래벨디자인 세부조정,  버튼가이드와 튜토리얼 새로만들기

+++ Collision.cpp
@@ -179,7 +179,7 @@ bool Collision::Check_Collision(const Triangle& tri, const ML::Vec3& p)
 
 	//誤差まで確認(誤差の範囲は変わる余地がある 2018/03/16)
 	//0.99756405026
-	if (check  >= _CMATH_::cos(ML::ToRadian(359)))
+	if (check  >= _CMATH_::cosf(ML::ToRadian(359)))
 	{
 		return true;
 	}
